FS-1087 fix application status (#16)
Fixes:
A questions' status could not be set to IN_PROGRESS if it contained a mix of complete and incomplete fields. FIXED
A section is set to completed even if a question within the section still has a status of not_started. FIXED
An application is set to completed even if a section within the application has a status of not_started. FIXED