WIP: moves patching logic down to the service layer
Moving the logic for updating applications into the code fetching the
application and forms lets us update a copy of each directly and then
run one single db commit transaction -- the code should read a bit more
simply and we should get the "all or nothing" guarantee of a
transaction.

Also expose various properties for the API the frontend relies on.